feat(earn): rename interest to earn
