+++ packages/blockchain-wallet-v4/src/network/api/earn/index.ts
