+++ packages/blockchain-wallet-v4/src/network/api/earn/types.ts
