+++ packages/blockchain-wallet-v4/src/types/index.ts
@@ -116,8 +116,8 @@ export * from '../coinSelection/types'
 export * from '../exchange/currencies'
 export * from '../network/api/buySell/types'
 export * from '../network/api/custodial/types'
+export * from '../network/api/earn/types'
 export * from '../network/api/eth/types'
-export * from '../network/api/interest/types'
 export * from '../network/api/kyc/types'
 export * from '../network/api/misc/types'
 export * from '../network/api/profile/types'
